added chnages to the user module

+++ module/Category/src/Category/Controller/CategoryController.php
@@ -21,6 +21,7 @@ class CategoryController extends AbstractRestfulJsonController{
         $categories = array_map(function($category){
             return $category->toArray();
         }, $categories);
+        $this->getResponse()->setStatusCode(200);
         return new JsonModel($categories);
     }
 
@@ -30,10 +31,19 @@ class CategoryController extends AbstractRestfulJsonController{
 
     public function get($id){   
         // Action used for GET requests with resource Id
-        $category = $this->getEntityManager()->getRepository('Category\Entity\Category')->find($id);
-        return new JsonModel(
-    		$category->toArray()
-    	);
+        if(!empty($id)) {
+            $category = $this->getEntityManager()->getRepository('Category\Entity\Category')->find($id);
+            if(empty($category)) {
+                $this->getResponse()->setStatusCode(404);
+                return new JsonModel(array("status" => "error", "data" => "Category not found"));
+            }
+            $this->getResponse()->setStatusCode(200);
+            return new JsonModel(
+        		$category->toArray()
+        	);
+        }
+        $this->getResponse()->setStatusCode(400);
+        return new JsonModel(array("status" => "error", "data" => "post data is required"));
     }
 
     public function create($data){
@@ -43,29 +53,45 @@ class CategoryController extends AbstractRestfulJsonController{
         
         $this->getEntityManager()->persist($category);
         $this->getEntityManager()->flush();
-        
+        $this->getResponse()->setStatusCode(200);
         return new JsonModel($category->toArray());
     }
 
     public function update($id, $data){
         // Action used for PUT requests
-        $category = $this->getEntityManager()->getRepository('Category\Entity\Category')->find($id);
-        $category->set($data);
-        $category->validate($this->em);
-        
-        $this->getEntityManager()->flush();
-        
-        return new JsonModel($category->toArray());
+        if(!empty($id) && !empty($data)) {
+            $category = $this->getEntityManager()->getRepository('Category\Entity\Category')->find($id);
+            if(empty($id)) {
+                $this->getResponse()->setStatusCode(404);
+                return new JsonModel(array("status" => "error", "data" => "Category not found"));
+            }
+            $category->set($data);
+            $category->validate($this->em);
+            
+            $this->getEntityManager()->flush();
+            $this->getResponse()->setStatusCode(200);
+            return new JsonModel($category->toArray());
+        }
+        $this->getResponse()->setStatusCode(400);
+        return new JsonModel(array("status" => "error", "data" => "post data is required"));
     }
 
     public function delete($id){
         // Action used for DELETE requests
-        $category = $this->getEntityManager()->getRepository('Category\Entity\Category')->find($id);
-        $this->getEntityManager()->remove($category);
-        
-        $this->getEntityManager()->flush();
-        
-        return new JsonModel($category->toArray());
+        if(!empty($id) && $this->identity()->getRole() != 'user') {
+            $category = $this->getEntityManager()->getRepository('Category\Entity\Category')->find($id);
+            if(empty($category)) {
+                $this->getResponse()->setStatusCode(404);
+                return new JsonModel(array("status" => "error", "data" => "Category not found"));
+            }
+            $this->getEntityManager()->remove($category);
+            
+            $this->getEntityManager()->flush();
+            
+            return new JsonModel($category->toArray());
+        }
+        $this->getResponse()->setStatusCode(400);
+        return new JsonModel(array("status" => "error", "data" => "get data is required"));
     }
 
 }
